Update file architecture
The files were reorganized into utils, page, and assets.  Utils are concerned with codes and files that has no direct connection with the system. Page folder contains all the page files. The name "Pages" was not used instead of "Page" as "Pages" is recognized by the streamlit as a default file. Assets contain all the images. CreateDB and OCRs are in the "utils" directory. app.py and output_gen are in the root directory.